# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vireolin API",
        "description": "API توثيق لتطبيق Vireolin",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user",
                "description": "Register new user with the provided information.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User data for register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Full Name User",
                                        "type": "string",
                                        "maxLength": 100,
                                        "minLength": 3,
                                        "example": "Ali Aljndy"
                                    },
                                    "email": {
                                        "description": "user email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmed@example.com"
                                    },
                                    "password": {
                                        "description": "password must have at least 8 characters including uppercase, lowercase, numbers, and symbols",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "Password123!"
                                    },
                                    "password_confirmation": {
                                        "description": "Password Confirmation",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "phone_number": {
                                        "description": "",
                                        "type": "string",
                                        "maxLength": 15,
                                        "minLength": 8,
                                        "example": "0998680361"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Registered Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User registered. Please check your email to verify your account."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "JWT token",
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "token-type": {
                                                    "description": "Token Type",
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failed Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error in server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User",
                "description": "Login User With Provided info.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Login Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "user email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmed@example.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "logged in successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged in Successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "JWT token",
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Ali"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "ahmed@example.com"
                                                        },
                                                        "phone_number": {
                                                            "type": "string",
                                                            "example": "0998680361"
                                                        },
                                                        "email_verified_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": null
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "roles": {
                                                    "description": "User Roles",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "customer"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials. Please try again."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failed validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error in server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout User",
                "description": "logout user and destroy the token.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "User Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Logged Out Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "failed"
                    },
                    "message": {
                        "type": "string",
                        "example": "Error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "example": []
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "failed"
                    },
                    "message": {
                        "type": "string",
                        "example": "failed validation please confirm the input"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}